requirements.txt
# Tools
hypothesis
python-dotenv
pytest
pytest-cov

# Formatting
black

# Packages
numpy
napari
gsw
